// apps/backend/src/qr-acceso/qr-resultado.enum.ts
export enum QrResultado {
  VALIDO = 'QR_VALIDO',
  NO_ENCONTRADO = 'QR_NO_ENCONTRADO',
  INACTIVO = 'QR_INACTIVO',
  EXPIRADO = 'QR_EXPIRADO',
  TIPO_INVALIDO = 'QR_TIPO_INVALIDO',
  PARTICIPANTE_NO_VALIDADO = 'PARTICIPANTE_NO_VALIDADO',
  DIA_NO_AUTORIZADO = 'DIA_NO_AUTORIZADO',
  ACCESO_YA_UTILIZADO = 'ACCESO_YA_UTILIZADO',
  FUERA_DE_FECHA = 'FUERA_DE_FECHA_DEL_EVENTO',
}